requirements.txt
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-Login>=0.5.0
Pillow>=8.3.1
numpy>=1.19.5
tensorflow-cpu>=2.6.0
opencv-python-headless>=4.5.3.56
python-dotenv>=0.19.0
Werkzeug>=2.0.1 